refactor: pages/common 내부의 상수파일 src/common 하위로 이동  #82
- pages/common 제거
